# SNRAY-1244: authorization performance improvements